Started separating scenes into individual files
I don't know Phaser best practices and I don't know how to message pass between scenes... in the meantime, made a state.js file that keeps the game so they can start other scenes. I'm hoping there's an easier way to get around this.
