+++ src/elements.js
@@ -1,0 +1,6 @@
+export default {
+  buttonGameStart: document.getElementById("game-start"),
+  buttonGameReplay: document.getElementById("game-replay"),
+  app404: document.getElementById("app-404"),
+  appEnd: document.getElementById("app-end")
+};
