+++ src/options.js
@@ -1,0 +1,8 @@
+export default {
+  gravity: 400,
+  jumpForce: 700,
+  platformSpeed: 4,
+  platformSpeedIncrement: 0,
+  windowHeight: 600,
+  windowWidth: 800
+};
