+++ src/state.js
@@ -1,0 +1,13 @@
+let game;
+
+export default {
+  getGame() {
+    return game;
+  },
+  setGame(newGame) {
+    game = newGame;
+  },
+  startGameScene(scene) {
+    game.scene.start(scene);
+  }
+};
